1、修复了查询学生所有任务列表信息的bug 2、更加完善了数据库结构 3、初步完成了查询学生未完成的试卷 4、初步完成了开始答题 5、增加了Mapping里有关试卷，答卷，答题的转换 6、增加了题目类型的常量类 7、初步完成了提交试卷的接口

+++ ms_stu_pro_320_server_new/ruoyi-manage/src/main/java/com/mashang/domain/entity/TestAnswer.java
@@ -22,7 +22,7 @@ public class TestAnswer extends BaseModel {
 
     @ApiModelProperty(value = "答卷id")
     @TableId(value = "test_answer_id", type = IdType.AUTO)
-    private Integer testAnswerId;
+    private Long testAnswerId;
 
     @ApiModelProperty(value = "试卷id")
     private Integer testId;
